requirements.txt
sqlalchemy
python-telegram-bot
telethon
pydantic==2.4.1
backoff
python-dotenv

aiofiles==23.1.0
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
certifi==2022.12.7
chardet==4.0.0
click==7.1.2
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.95.2
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.12.0
httptools==0.1.2
httpx==0.24.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==2.0.1
orjson==3.5.2
promise==2.3
pydantic==1.8.2
python-dotenv==0.17.1
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.25.1
Rx==1.6.1
starlette==0.27.0
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.5
uvicorn==0.13.4
uvloop==0.14.0
watchgod==0.7
websockets==10.0
